refactor: split widget wallet integration base hooks
