+++ typescript/widgets/src/walletIntegrations/activeChains.tsx
@@ -3,14 +3,14 @@ import { useMemo } from 'react';
 import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '@hyperlane-xyz/sdk/providers/ConfiguredMultiProtocolProvider';
 import { type KnownProtocolType, ProtocolType } from '@hyperlane-xyz/utils';
 
-import { useAleoActiveChain } from './aleo.js';
-import { useCosmosActiveChain } from './cosmos.js';
-import { useEthereumActiveChain } from './ethereum.js';
-import { useRadixActiveChain } from './radix.js';
-import { useSolanaActiveChain } from './solana.js';
-import { useStarknetActiveChain } from './starknet.js';
+import { useAleoActiveChain } from './aleoBase.js';
+import { useCosmosActiveChain } from './cosmosBase.js';
+import { useEthereumActiveChain } from './ethereumBase.js';
+import { useRadixActiveChain } from './radixBase.js';
+import { useSolanaActiveChain } from './solanaBase.js';
+import { useStarknetActiveChain } from './starknetBase.js';
 import { type ActiveChainInfo } from './types.js';
-import { useTronActiveChain } from './tron.js';
+import { useTronActiveChain } from './tronBase.js';
 
 export function useActiveChains(multiProvider: MultiProtocolProvider): {
   chains: Record<KnownProtocolType, ActiveChainInfo>;
