+++ typescript/widgets/src/walletIntegrations/connectFns.tsx
@@ -2,13 +2,13 @@ import { useMemo } from 'react';
 
 import { type KnownProtocolType, ProtocolType } from '@hyperlane-xyz/utils';
 
-import { useAleoConnectFn } from './aleo.js';
-import { useCosmosConnectFn } from './cosmos.js';
-import { useEthereumConnectFn } from './ethereum.js';
-import { useRadixConnectFn } from './radix.js';
-import { useSolanaConnectFn } from './solana.js';
-import { useStarknetConnectFn } from './starknet.js';
-import { useTronConnectFn } from './tron.js';
+import { useAleoConnectFn } from './aleoBase.js';
+import { useCosmosConnectFn } from './cosmosBase.js';
+import { useEthereumConnectFn } from './ethereumBase.js';
+import { useRadixConnectFn } from './radixBase.js';
+import { useSolanaConnectFn } from './solanaBase.js';
+import { useStarknetConnectFn } from './starknetBase.js';
+import { useTronConnectFn } from './tronBase.js';
 
 export function useConnectFns(): Record<KnownProtocolType, () => void> {
   const onConnectEthereum = useEthereumConnectFn();
