+++ typescript/widgets/src/walletIntegrations/solana.ts
@@ -1,7 +1,6 @@
-import { useConnection, useWallet } from '@solana/wallet-adapter-react';
-import { useWalletModal } from '@solana/wallet-adapter-react-ui';
+import { useWallet } from '@solana/wallet-adapter-react';
 import { Connection } from '@solana/web3.js';
-import { useCallback, useMemo } from 'react';
+import { useCallback } from 'react';
 
 import {
   ProviderType,
@@ -11,82 +10,23 @@ import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '@
 import type { ITokenMetadata } from '@hyperlane-xyz/sdk/token/ITokenMetadata';
 import type { ChainName } from '@hyperlane-xyz/sdk/types';
 import type { WarpTypedTransaction } from '@hyperlane-xyz/sdk/warp/types';
-import { ProtocolType } from '@hyperlane-xyz/utils';
 
 import { widgetLogger } from '../logger.js';
 
 import {
-  AccountInfo,
-  ActiveChainInfo,
   ChainTransactionFns,
   SwitchNetworkFns,
-  WalletDetails,
   WatchAssetFns,
 } from './types.js';
-import { findChainByRpcUrl } from './utils.js';
 
 const logger = widgetLogger.child({ module: 'walletIntegrations/solana' });
-
-export function useSolanaAccount(
-  _multiProvider: MultiProtocolProvider,
-): AccountInfo {
-  const { publicKey, connected, wallet } = useWallet();
-  const isReady = !!(publicKey && wallet && connected);
-  const address = publicKey?.toBase58();
-
-  return useMemo<AccountInfo>(
-    () => ({
-      protocol: ProtocolType.Sealevel,
-      addresses: address ? [{ address: address }] : [],
-      isReady: isReady,
-    }),
-    [address, isReady],
-  );
-}
-
-export function useSolanaWalletDetails() {
-  const { wallet } = useWallet();
-  const { name, icon } = wallet?.adapter || {};
-
-  return useMemo<WalletDetails>(
-    () => ({
-      name,
-      logoUrl: icon,
-    }),
-    [name, icon],
-  );
-}
-
-export function useSolanaConnectFn(): () => void {
-  const { setVisible } = useWalletModal();
-  return useCallback(() => setVisible(true), [setVisible]);
-}
-
-export function useSolanaDisconnectFn(): () => Promise<void> {
-  const { disconnect } = useWallet();
-  return disconnect;
-}
-
-export function useSolanaActiveChain(
-  multiProvider: MultiProtocolProvider,
-): ActiveChainInfo {
-  const { connection } = useConnection();
-  const connectionEndpoint = connection?.rpcEndpoint;
-  return useMemo<ActiveChainInfo>(() => {
-    try {
-      const hostname = new URL(connectionEndpoint).hostname;
-      const metadata = findChainByRpcUrl(multiProvider, hostname);
-      if (!metadata) return {};
-      return {
-        chainDisplayName: metadata.displayName,
-        chainName: metadata.name,
-      };
-    } catch (error) {
-      logger.warn('Error finding sol active chain', error);
-      return {};
-    }
-  }, [connectionEndpoint, multiProvider]);
-}
+export {
+  useSolanaAccount,
+  useSolanaActiveChain,
+  useSolanaConnectFn,
+  useSolanaDisconnectFn,
+  useSolanaWalletDetails,
+} from './solanaBase.js';
 
 export function useSolanaSwitchNetwork(): SwitchNetworkFns {
   const onSwitchNetwork = useCallback(async (chainName: ChainName) => {
